+++ tests/checkpointing/test_model_checkpoint.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import logging
 import math
 import os
 import pickle
@@ -346,9 +347,9 @@ class ModelCheckpointTestInvocations(ModelCheckpoint):
     def on_train_start(self, trainer, pl_module):
         torch.save = Mock(wraps=torch.save)
 
-    def on_save_checkpoint(self, trainer, pl_module):
+    def on_save_checkpoint(self, trainer, pl_module, checkpoint):
         # expect all ranks to run but only rank 0 will actually write the checkpoint file
-        super().on_save_checkpoint(trainer, pl_module)
+        super().on_save_checkpoint(trainer, pl_module, checkpoint)
         self.on_save_checkpoint_count += 1
 
     def on_train_end(self, trainer, pl_module):
@@ -677,7 +678,8 @@ def test_model_checkpoint_save_last_warning(tmpdir, caplog, max_epochs, should_v
         callbacks=[ModelCheckpoint(monitor='early_stop_on', dirpath=tmpdir, save_top_k=0, save_last=save_last)],
         max_epochs=max_epochs,
     )
-    trainer.fit(model)
+    with caplog.at_level(logging.INFO):
+        trainer.fit(model)
     assert caplog.messages.count('Saving latest checkpoint...') == save_last
 
 
@@ -898,16 +900,6 @@ def test_configure_model_checkpoint(tmpdir):
     trainer = Trainer(callbacks=[callback1, callback2], **kwargs)
     assert trainer.checkpoint_callback == callback1
     assert trainer.checkpoint_callbacks == [callback1, callback2]
-
-    with pytest.warns(DeprecationWarning, match='will no longer be supported in v1.3'):
-        trainer = Trainer(checkpoint_callback=callback1, **kwargs)
-        assert [c for c in trainer.callbacks if isinstance(c, ModelCheckpoint)] == [callback1]
-        assert trainer.checkpoint_callback == callback1
-
-    with pytest.warns(DeprecationWarning, match="will no longer be supported in v1.3"):
-        trainer = Trainer(checkpoint_callback=callback1, callbacks=[callback2], **kwargs)
-        assert trainer.checkpoint_callback == callback2
-        assert trainer.checkpoint_callbacks == [callback2, callback1]
 
     with pytest.raises(MisconfigurationException, match="checkpoint_callback=False but found ModelCheckpoint"):
         Trainer(checkpoint_callback=False, callbacks=[callback1], **kwargs)