+++ tests/models/test_horovod.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 import json
 import os
-import platform
 import shlex
 import subprocess
 import sys
@@ -29,9 +28,10 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.accelerators import CPUAccelerator
 from pytorch_lightning.metrics.classification.accuracy import Accuracy
 from pytorch_lightning.trainer.states import TrainerState
-from pytorch_lightning.utilities import _APEX_AVAILABLE, _HOROVOD_AVAILABLE, _NATIVE_AMP_AVAILABLE
+from pytorch_lightning.utilities import _HOROVOD_AVAILABLE
 from tests.helpers import BoringModel
 from tests.helpers.advanced_models import BasicGAN
+from tests.helpers.runif import RunIf
 
 if _HOROVOD_AVAILABLE:
     import horovod
@@ -66,7 +66,7 @@ def _run_horovod(trainer_options, on_gpu=False):
     assert exit_code == 0
 
 
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
+@RunIf(skip_windows=True)
 def test_horovod_cpu(tmpdir):
     """Test Horovod running multi-process on CPU."""
     trainer_options = dict(
@@ -83,7 +83,7 @@ def test_horovod_cpu(tmpdir):
     _run_horovod(trainer_options)
 
 
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
+@RunIf(skip_windows=True)
 def test_horovod_cpu_implicit(tmpdir):
     """Test Horovod without specifying a backend, inferring from env set by `horovodrun`."""
     trainer_options = dict(
@@ -99,9 +99,8 @@ def test_horovod_cpu_implicit(tmpdir):
     _run_horovod(trainer_options)
 
 
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
 @pytest.mark.skipif(not _HOROVOD_NCCL_AVAILABLE, reason="test requires Horovod with NCCL support")
-@pytest.mark.skipif(torch.cuda.device_count() < 2, reason="test requires multi-GPU machine")
+@RunIf(min_gpus=2, skip_windows=True)
 def test_horovod_multi_gpu(tmpdir):
     """Test Horovod with multi-GPU support."""
     trainer_options = dict(
@@ -120,10 +119,8 @@ def test_horovod_multi_gpu(tmpdir):
 
 
 @pytest.mark.skip(reason="Horovod has a problem with broadcast when using apex?")
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
 @pytest.mark.skipif(not _HOROVOD_NCCL_AVAILABLE, reason="test requires Horovod with NCCL support")
-@pytest.mark.skipif(torch.cuda.device_count() < 2, reason="test requires multi-GPU machine")
-@pytest.mark.skipif(not _APEX_AVAILABLE, reason="test requires apex")
+@RunIf(min_gpus=2, skip_windows=True, amp_apex=True)
 def test_horovod_apex(tmpdir):
     """Test Horovod with multi-GPU support using apex amp."""
     trainer_options = dict(
@@ -144,10 +141,8 @@ def test_horovod_apex(tmpdir):
 
 
 @pytest.mark.skip(reason="Skip till Horovod fixes integration with Native torch.cuda.amp")
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
 @pytest.mark.skipif(not _HOROVOD_NCCL_AVAILABLE, reason="test requires Horovod with NCCL support")
-@pytest.mark.skipif(torch.cuda.device_count() < 2, reason="test requires multi-GPU machine")
-@pytest.mark.skipif(not _NATIVE_AMP_AVAILABLE, reason="test requires torch.cuda.amp")
+@RunIf(min_gpus=2, skip_windows=True, amp_native=True)
 def test_horovod_amp(tmpdir):
     """Test Horovod with multi-GPU support using native amp."""
     trainer_options = dict(
@@ -167,9 +162,8 @@ def test_horovod_amp(tmpdir):
     _run_horovod(trainer_options, on_gpu=True)
 
 
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
 @pytest.mark.skipif(not _HOROVOD_NCCL_AVAILABLE, reason="test requires Horovod with NCCL support")
-@pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires GPU machine")
+@RunIf(min_gpus=1, skip_windows=True)
 def test_horovod_transfer_batch_to_gpu(tmpdir):
 
     class TestTrainingStepModel(BoringModel):
@@ -197,7 +191,7 @@ def test_horovod_transfer_batch_to_gpu(tmpdir):
     tpipes.run_model_test_without_loggers(trainer_options, model)
 
 
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
+@RunIf(skip_windows=True)
 def test_horovod_multi_optimizer(tmpdir):
     model = BasicGAN()
 
@@ -232,7 +226,7 @@ def test_horovod_multi_optimizer(tmpdir):
 # TODO: unclear Horovod failure...
 @pytest.mark.skip(reason="unclear Horovod failure...")
 @pytest.mark.skipif(not _HOROVOD_AVAILABLE, reason="Horovod is unavailable")
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
+@RunIf(skip_windows=True)
 def test_result_reduce_horovod(tmpdir):
     """Make sure result logging works with Horovod.
 
@@ -283,7 +277,7 @@ def test_result_reduce_horovod(tmpdir):
 # TODO: unclear Horovod failure...
 @pytest.mark.skip(reason="unclear Horovod failure...")
 @pytest.mark.skipif(not _HOROVOD_AVAILABLE, reason="Horovod is unavailable")
-@pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
+@RunIf(skip_windows=True)
 def test_accuracy_metric_horovod():
     num_batches = 10
     batch_size = 16
@@ -334,7 +328,7 @@ def test_accuracy_metric_horovod():
     horovod.run(_compute_batch, np=2)
 
 
-# @pytest.mark.skipif(platform.system() == "Windows", reason="Horovod is not supported on Windows")
+# @RunIf(skip_windows=True)
 # def test_horovod_multi_optimizer_with_scheduling_stepping(tmpdir):
 #     model = BoringModel()
 #     model.configure_optimizers = model.configure_optimizers__multiple_schedulers